+++ io_u.c
@@ -314,8 +314,7 @@ static int get_next_block(struct thread_data *td, struct io_u *io_u,
 		else if (b != -1ULL)
 			io_u->offset = b * td->o.ba[ddir];
 		else {
-			log_err("fio: bug in offset generation: offset=%llu, b=%llu\n",
-								offset, b);
+			log_err("fio: bug in offset generation: offset=%llu, b=%llu\n", (unsigned long long) offset, (unsigned long long) b);
 			ret = 1;
 		}
 	}
